pyproject.toml
[tool.poetry]
name = "slitherio-ml"
version = "0.1.0"
description = ""
authors = ["Carlos Kidman"]

[tool.poetry.dependencies]
python = "^3.8"
tensorflow = "^2.3.1"
pyleniumio = "^1.11.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
